Revert "Fix conversion to from float to int"
This reverts commit f5d003d1e21ffe26b80881ae6996b1ab76f1b4dc.

We decided to postpone this change until a later release, in maybe a
different shape.

+++ internal/compiler/builtin_macros.rs
@@ -249,7 +249,7 @@ fn rgb_macro(
                         op: '*',
                     }
                 } else {
-                    expr.maybe_convert_to(Type::Float32, &n, diag)
+                    expr.maybe_convert_to(Type::Int32, &n, diag)
                 }
             } else {
                 expr.maybe_convert_to(Type::Float32, &n, diag)
